+++ server/config/index.mjs
@@ -45,6 +45,14 @@ dotenv.config();
 
 const nodeEnv = process.env.NODE_ENV || 'development';
 
+// Check requirement for minimum node version
+const minNodeVersion = 16;
+if (parseInt(process.version.replace('v', '').split('.')[0]) < minNodeVersion) {
+  console.error('Error: this program requires node version ' +
+    minNodeVersion.toString() + ' or greater.');
+  process.exit(1);
+}
+
 let credentials = null;
 try {
   credentials = JSON.parse(fs.readFileSync('./credentials.json', 'utf8'));